+++ bitClassroom/app/controllers/posts/PostController.java
@@ -1,7 +1,6 @@
 package controllers.posts;
 
 import helpers.Authorization;
-import helpers.CurrentUserFilter;
 import helpers.SessionHelper;
 import models.Post;
 import models.user.User;
@@ -9,10 +8,9 @@ import org.apache.commons.io.FileUtils;
 import org.joda.time.DateTime;
 import play.Logger;
 import play.Play;
-import play.mvc.Http.MultipartFormData;
-import play.mvc.Http.MultipartFormData.FilePart;
 import play.data.Form;
 import play.mvc.Controller;
+import play.mvc.Http.MultipartFormData;
 import play.mvc.Result;
 import play.mvc.Security;
 import views.html.posts.newpost;
@@ -20,7 +18,6 @@ import views.html.posts.newpost;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 /**
