+++ bitClassroom/app/controllers/users/AdminController.java
@@ -4,6 +4,7 @@ import com.avaje.ebean.Ebean;
 import helpers.Authorization;
 import helpers.SessionHelper;
 import models.ErrorLog;
+import models.course.Course;
 import models.report.Field;
 import models.user.Role;
 import models.user.User;
@@ -21,6 +22,7 @@ import views.html.admins.allerrors;
 import views.html.admins.userlist;
 import views.html.dailyreports.dailyraport;
 import views.html.admins.setingsdailyraport;
+import views.html.fillClassDetails;
 
 
 import javax.persistence.PersistenceException;
@@ -33,6 +35,9 @@ import java.util.List;
 public class AdminController extends Controller {
 
     private final Form<User> userForm = Form.form(User.class);
+
+    private final Form<Field> fieldForm = Form.form(Field.class);
+
 
 
     /**
@@ -57,7 +62,6 @@ public class AdminController extends Controller {
 
     public Result listOfUser(){
         return ok(userlist.render(User.findAll()));
-
     }
 
     /**
@@ -161,10 +165,34 @@ public class AdminController extends Controller {
         return ok(setingsdailyraport.render());
     }
 
-    private static final Form<Field> fieldForm = Form.form(Field.class);
 
 
 
     public Result saveField() {
-        return ok(dailyraport.render(fieldForm)); }
+
+        return ok("ad");
+    }
+
+
+    private final Form<Course> courseForm = Form.form(Course.class);
+
+    public Result addCourse() {
+        return ok(fillClassDetails.render(courseForm));
+    }
+
+    public Result saveCourse() {
+        Form<Course> boundForm = courseForm.bindFromRequest();
+
+        String name = boundForm.bindFromRequest().field("name").value();
+        String description = boundForm.bindFromRequest().field("description").value();
+        String teacher = boundForm.bindFromRequest().field("teacher").value();
+
+        Course course = new Course(name, description, teacher);
+        course.save();
+        flash("success", "You successfully added new course.");
+        return redirect("/index"); // TODO add call to route for listing posts
+
+
+    }
+
 }
