+++ bitClassroom/app/views/index.scala.html
@@ -43,7 +43,7 @@
     }
     @if(user != null) {
 
-        <a id="add" href="@controllers.course.routes.CourseController.addCourse()" class="btn btn-info">
+        <a id="add" href="@controllers.users.routes.AdminController.addCourse()" class="btn btn-info">
             Add class</a>
