+++ bitClassroom/conf/evolutions/default/1.sql
@@ -3,8 +3,20 @@
 
 # --- !Ups
 
+create table course (
+  id                        bigint auto_increment not null,
+  name                      varchar(100),
+  description               varchar(100),
+  teacher                   varchar(100),
+  create_date               datetime,
+  created_by                varchar(255),
+  update_date               datetime,
+  updated_by                varchar(255),
+  constraint uq_course_name unique (name),
+  constraint pk_course primary key (id))
+;
+
 create table daily_report (
-<<<<<<< HEAD
   id                        bigint auto_increment not null,
   name                      varchar(255),
   create_date               datetime,
@@ -13,21 +25,7 @@ create table daily_report (
 ;
 
 create table error_log (
-=======
-<<<<<<< HEAD
-  id                        bigint not null,
-=======
->>>>>>> master
   id                        bigint auto_increment not null,
->>>>>>> master
-  name                      varchar(255),
-  create_date               datetime,
-  data                      varchar(255),
-  constraint pk_daily_report primary key (id))
-;
-
-create table error_log (
-  id                        bigint not null,
   message                   varchar(255),
   loged_date                datetime,
   constraint pk_error_log primary key (id))
@@ -40,11 +38,11 @@ create table field (
 ;
 
 create table post (
-  id                        bigint not null,
+  id                        bigint auto_increment not null,
   title                     varchar(255),
   content                   text,
   post_type                 integer,
-  visible_mentors           boolean,
+  visible_mentors           tinyint(1) default 0,
   date                      varchar(255),
   link                      varchar(255),
   user_id                   bigint,
@@ -60,13 +58,13 @@ create table report_field (
 ;
 
 create table role (
-  id                        bigint not null,
+  id                        bigint auto_increment not null,
   name                      varchar(10),
   constraint pk_role primary key (id))
 ;
 
 create table user (
-  id                        bigint not null,
+  id                        bigint auto_increment not null,
   email                     varchar(100),
   password                  varchar(100),
   first_name                varchar(100),
@@ -93,21 +91,17 @@ create table user (
 ;
 
 
+create table course_user (
+  course_id                      bigint not null,
+  user_id                        bigint not null,
+  constraint pk_course_user primary key (course_id, user_id))
+;
+
 create table user_role (
   user_id                        bigint not null,
   role_id                        bigint not null,
   constraint pk_user_role primary key (user_id, role_id))
 ;
-create sequence daily_report_seq;
-
-create sequence error_log_seq;
-
-create sequence post_seq;
-
-create sequence role_seq;
-
-create sequence user_seq;
-
 alter table post add constraint fk_post_user_1 foreign key (user_id) references user (id) on delete restrict on update restrict;
 create index ix_post_user_1 on post (user_id);
 alter table report_field add constraint fk_report_field_dailyReport_2 foreign key (daly_id) references daily_report (id) on delete restrict on update restrict;
@@ -117,37 +111,26 @@ create index ix_report_field_field_3 on report_field (field_id);
 
 
 
+alter table course_user add constraint fk_course_user_course_01 foreign key (course_id) references course (id) on delete restrict on update restrict;
+
+alter table course_user add constraint fk_course_user_user_02 foreign key (user_id) references user (id) on delete restrict on update restrict;
+
 alter table user_role add constraint fk_user_role_user_01 foreign key (user_id) references user (id) on delete restrict on update restrict;
 
 alter table user_role add constraint fk_user_role_role_02 foreign key (role_id) references role (id) on delete restrict on update restrict;
 
 # --- !Downs
 
-SET REFERENTIAL_INTEGRITY FALSE;
+SET FOREIGN_KEY_CHECKS=0;
 
-drop table if exists daily_report;
+drop table course;
 
-drop table if exists error_log;
-
-drop table if exists post;
-
-drop table if exists role;
-
-drop table if exists user;
-
-drop table if exists user_role;
-
-<<<<<<< HEAD
-SET REFERENTIAL_INTEGRITY TRUE;
-=======
-drop table daily_report;
+drop table course_user;
 
 drop table daily_report;
 
 drop table error_log;
->>>>>>> master
 
-<<<<<<< HEAD
 drop table field;
 
 drop table post;
@@ -155,15 +138,10 @@ drop table post;
 drop table report_field;
 
 drop table role;
-=======
-drop sequence if exists daily_report_seq;
 
-drop sequence if exists error_log_seq;
->>>>>>> master
+drop table user;
 
-drop sequence if exists post_seq;
+drop table user_role;
 
-drop sequence if exists role_seq;
-
-drop sequence if exists user_seq;
+SET FOREIGN_KEY_CHECKS=1;
